requirements.txt
requests
bs4
lxml
pytelegrambotapi
python-decouple